refactor(ui): split shared primitives and pickers
Move shared atoms, icons, overlays, chip input, floating dropdown style logic, and provider/model/agent picker components into dedicated files. Keep shared/ui.tsx as the compatibility barrel for existing imports.
